+++ spec/classes/incron_spec.rb
@@ -180,13 +180,11 @@ describe 'incron' do
         }
       end
 
-      # FIXME
-      # rspec-puppet-facts `on_supported_os` doesn't do 18.04 yet
-      %w[14.04 16.04 18.04].each do |os_ver|
+      on_supported_os.each do |os_ver, facts|
         context "on Ubuntu #{os_ver}" do
-          let(:facts) { { os: { release: { full: os_ver } } } }
+          let(:facts) { facts }
 
-          if os_ver == '14.04'
+          if facts[:os]['release']['full'] == '14.04'
             it { is_expected.not_to contain_service('incron') }
           else
             it { is_expected.to contain_service('incron').with_ensure(:stopped).with_provider(:systemd) }